+++ ocaml-lsp-server/src/ocaml_lsp_server.mli
@@ -1,4 +1,4 @@
-val run : Lsp.Cli.Channel.t -> read_dot_merlin:bool -> unit -> unit
+val run : Lsp.Cli.Channel.t -> unit -> unit
 
 module Diagnostics = Diagnostics
 module Version = Version